coco-conversion/requirements.txt
Pillow
pandas
pyarrow
tqdm
